+++ src/nz/ac/vuw/ecs/swen225/gp22/app/Point.java
@@ -1,18 +1,24 @@
 package nz.ac.vuw.ecs.swen225.gp22.app;
 
 import java.util.Random;
-record Point(double x, double y){
-    //Note: x==with, y==height
-    public Point add(double x,double y){
+
+/**
+ * Point class for representing points in the game board
+ * @author Kevin Zeng
+ * ID: 300563468
+ */
+record Point(int x, int y){
+    //Note: x==width, y==height
+    public Point add(int x,int y){
         return new Point(x()+x, y()+y);
     }
     public Point add(Point p){
         return add(p.x, p.y);
     }
-    public Point times(double x, double y) {
+    public Point times(int x, int y) {
         return new Point(x()*x, y()*y);
     }
-    public Point times(double v) {
+    public Point times(int v) {
         return new Point(x()*v, y()*v);
     }
 
@@ -24,6 +30,15 @@ record Point(double x, double y){
         return Math.sqrt(x*x+y*y);
     }
 
+    /**
+     * Generates a random point inside the rectangle formed by
+     * (minX,minY) and (maxX,maxY)
+     * @param minX minimum x value
+     * @param maxX maximum x value
+     * @param minY minimum y value
+     * @param maxY maximum y value
+     * @return A new Point object with random coordinates inside the bounds
+     */
     public static Point randomPoint(int minX, int maxX, int minY, int maxY) {
         Random random = new Random();
         return new Point(random.nextInt(maxX - minX) + minX,random.nextInt(maxY - minY) + minY);
